Refactor: move test files out from 'tests' folder
unnecessary nesting
